+++ assets/css/global.scss
@@ -478,10 +478,6 @@ legend {
   padding-right: 0px;
 }
 
-.sidebar {
-  height: calc(100vh - 100px);
-}
-
 .mx-datepicker-popup {
   z-index: 2000 !important;;
 }
