+++ pages/chitchat/_id.vue
@@ -2,10 +2,7 @@
   <div v-if="me">
     <b-row class="m-0">
       <b-col cols="0" md="3" class="d-none d-md-block">
-        <div class="d-flex flex-column sidebar">
-          <CommunityEventSidebar class="justify-content-start flex-grow-1" style="overflow-y: auto" />
-          <BotLeftBox class="justify-content-end flex-shrink-2" />
-        </div>
+        <sidebar-left />
       </b-col>
       <b-col cols="12" md="6" class="newsfeedHolder p-0">
         <b-card v-if="!id">
@@ -157,10 +154,7 @@
         </div>
       </b-col>
       <b-col cols="0" md="3" class="d-none d-md-block">
-        <div class="d-flex flex-column sidebar">
-          <VolunteerOpportunitySidebar class="justify-content-start flex-grow-1" style="overflow-y: auto" />
-          Job ads go here
-        </div>
+        <sidebar-right />
       </b-col>
     </b-row>
   </div>
@@ -183,20 +177,16 @@
 import loginRequired from '@/mixins/loginRequired.js'
 import twem from '~/assets/js/twem'
 import NewsThread from '~/components/NewsThread.vue'
-const CommunityEventSidebar = () =>
-  import('../../components/CommunityEventSidebar')
-const VolunteerOpportunitySidebar = () =>
-  import('../../components/VolunteerOpportunitySidebar')
 const OurFilePond = () => import('~/components/OurFilePond')
-const BotLeftBox = () => import('~/components/BotLeftBox')
+const SidebarLeft = () => import('~/components/sidebar-left.vue')
+const SidebarRight = () => import('~/components/sidebar-right.vue')
 
 export default {
   components: {
-    VolunteerOpportunitySidebar,
-    CommunityEventSidebar,
     NewsThread,
     OurFilePond,
-    BotLeftBox
+    SidebarLeft,
+    SidebarRight
   },
   mixins: [loginRequired],
